Cors, claim model, event model, reservation model

+++ api/controllers/users.controller.js
@@ -1,6 +1,7 @@
 const User = require("../models/user.model");
 const createError = require("http-errors");
 const mailer = require("../config/mailer.config");
+const jwt = require('jsonwebtoken');
 
 
 module.exports.list = (req, res, next) => {
@@ -33,6 +34,10 @@ module.exports.confirm = (req, res, next) => {
 module.exports.detail = (req, res, next) => res.json(req.user);
 
 module.exports.update = (req, res, next) => {
+  if (req.user.id !== req.params.id){
+    return next(createError(403, "Forbidden"));
+  }
+
   Object.assign(req.user, req.body);
   req.user
     .save()
@@ -41,6 +46,11 @@ module.exports.update = (req, res, next) => {
 };
 
 module.exports.delete = (req, res, next) => {
+  if (req.user.id !== req.params.id){
+    return next(createError(403, "Forbidden"));
+  }
+
+
   User.deleteOne({ _id: req.user.id })
     .then(() => res.status(204).send())
     .catch(next);
@@ -49,8 +59,12 @@ module.exports.delete = (req, res, next) => {
 module.exports.login = (req, res, next) => {
   User.findOne({ email: req.body.email })
     .then((user) => {
-      if (!user) {
+      if (!user || !req.body.password) {
         return next(createError(401, "Invalid credentials"));
+      }
+
+      if (!user.confirm) {
+        return next(createError(401, "Please, confirm your account"))
       }
 
       user.checkPassword(req.body.password)
@@ -58,6 +72,8 @@ module.exports.login = (req, res, next) => {
           if (!match) {
             return next(createError(401, "Invalid credentials"));
           }
+
+          const token = jwt.sign({ sub: user.id, exp: Date.now() / 1000 +3_600 }, process.env.JWT_SECRET)
 
           res.json(user);
         });
